Move form to reusable component
